backend/requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# OCR
paddlepaddle==2.6.2
paddleocr==2.7.3

# AI/LLM
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
Pillow==10.2.0
pdf2image==1.17.0

# Documentation
pdoc==14.4.0
email-validator==2.1.1
